feat(server,ent): GraphQL endpoint via entgql + gqlgen
Wires `/graphql` into the Echo server with a Relay-style schema
generated from the ent graph. The catalog read surface (Recording,
Show, Performer, CollectionEntry, WantsEntry, SyncRun, plus the
Recording → cast_entries / versions edges) is reachable through
cursor-paginated connections + entgql-generated where/order inputs.

- Annotates the in-scope ent schemas with entgql.RelayConnection +
  entgql.QueryField + per-field OrderField; out-of-scope entities
  (Profile, JobRun, JobState, HistoryEvent, ManualImportQueue,
  Character, RecordingImageChoice) carry entgql.Skip(SkipAll) so
  their non-int64 PKs do not collide with the relay schema's
  mixed-id-types check.
- Promotes RecordingVersion / SyncRun / CastEntry auto-increment
  IDs from int → int64 so every node type shares a uniform Go ID
  type — the SQLite column representation is unchanged.
- Adds typed `recording(id:)` / `show(id:)` / `performer(id:)` /
  `collectionEntry(id:)` / `wantsEntry(id:)` query shortcuts in
  schema.graphql. The Relay node(id:) resolver returns a documented
  error because promptbook does not maintain ent's `ent_types`
  meta table (we own migrations via goose, not ent's WithGlobalUniqueID).
- Mounts POST/GET /graphql + GET /graphql/playground on Echo before
  the SPA catch-all so route ordering is explicit.
- Five smoke tests against the fixture-backed server cover the
  playground, recording(id:), recordings connection w/ orderBy +
  pagination, shows connection drilling into the recordings edge,
  and the node-unsupported error path.

`go generate ./internal/...` regenerates ent + ent.graphql + gqlgen
artifacts in one pass; `task generate` is the make-friendly entry
point.

+++ internal/ent/castentry/where.go
@@ -9,47 +9,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.CastEntry {
+func ID(id int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.CastEntry {
+func IDEQ(id int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.CastEntry {
+func IDNEQ(id int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.CastEntry {
+func IDIn(ids ...int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.CastEntry {
+func IDNotIn(ids ...int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.CastEntry {
+func IDGT(id int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.CastEntry {
+func IDGTE(id int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.CastEntry {
+func IDLT(id int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.CastEntry {
+func IDLTE(id int64) predicate.CastEntry {
 	return predicate.CastEntry(sql.FieldLTE(FieldID, id))
 }
