+++ internal/ent/syncrun/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.SyncRun {
+func ID(id int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.SyncRun {
+func IDEQ(id int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.SyncRun {
+func IDNEQ(id int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.SyncRun {
+func IDIn(ids ...int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.SyncRun {
+func IDNotIn(ids ...int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.SyncRun {
+func IDGT(id int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.SyncRun {
+func IDGTE(id int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.SyncRun {
+func IDLT(id int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.SyncRun {
+func IDLTE(id int64) predicate.SyncRun {
 	return predicate.SyncRun(sql.FieldLTE(FieldID, id))
 }
